Moves transform component to component package
