+++ src/pyrite/transform.py
@@ -1,7 +1,7 @@
 import pyrite._transform.transform
-import pyrite._transform.transform_component
+import pyrite._component.transform_component
 import pyrite._transform.transform_system
 
 Transform = pyrite._transform.transform.Transform
-TransformComponent = pyrite._transform.transform_component.TransformComponent
+TransformComponent = pyrite._component.transform_component.TransformComponent
 TransformSystem = pyrite._transform.transform_system.TransformSystem
